ref: Move button to styled folder in common
