+++ src/modules/cabinet/components/Address/CabinetAddressForm.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Field, Form, Formik } from 'formik';
 import * as Styled from './CabinetAddressForm.styled';
-import { Button as StyledButton } from '../../../common/button/button.styled';
+import { Button as StyledButton } from '../../../common/styled/button.styled';
 import { addressValidationSchema } from '../../../../schemas/address.schema';
 import { IAddNewAddress, IAddress, IUpdateAddress } from '../../../../types/address';
 import addressService from '../../../../services/address.service';
