+++ src/modules/catalogue/pages/product-card/ProductCardMainInfo.tsx
@@ -1,7 +1,7 @@
 import React, { ChangeEventHandler, useContext, useState } from 'react';
 import RatingBar from '../../../common/ratingBar/RatingBar';
 import GoBack from '../../../common/goBack/GoBack';
-import * as Styled from '../../../common/button/button.styled';
+import * as Styled from '../../../common/styled/button.styled';
 import { IBasketItemQuantity } from '../../../../services/basket.service';
 import {
   IProductAlt,
